refactor(core): extract appendJitContextToParts util for non-string content
Move the logic for wrapping non-string content (images, PDFs) with JIT
context into a reusable appendJitContextToParts function in jit-context.ts.
This simplifies read-file.ts and makes the util available for other tools.
